cgs-mvp/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24,<0.26
supabase==2.3.4
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.4.0
slowapi==0.1.9
